+++ CHANGES.rst
@@ -17,6 +17,8 @@ Version 3.x.x
 - Accept :class:`uuid.UUID` instances in the :class:`~validators.UUID` validator,
   and translate unexpected types into :class:`~validators.ValidationError`
   instead of letting :class:`TypeError` escape. :issue:`549` :pr:`769`
+- Escape double quotes in :class:`~markupsafe.Markup` attribute values
+  passed to :func:`~widgets.html_params`. :issue:`801`
 
 Version 3.2.2
 -------------